[SDDK] rebuild SDAPI requests on each retry
Fixes a panic, Reqwest's builder doesn't support copying the body out of
a request, the entire thing is moved out of the builder when a request
is built. As a result the builder can't be told to build a request twice
as that would require allocating and rebuilding it twice. So instead we
will just move all that logic inside the loop, it's not expensive anyway